requirements.txt
cycler==0.9.0
Django==1.9
matplotlib==1.5.0
numpy==1.10.1
pyparsing==2.0.6
python-dateutil==2.4.2
python-Levenshtein==0.12.0
pytz==2015.7
six==1.10.0
